refactor: Move views folder from src/view/components to src/view
