+++ src/view/mod.rs
@@ -1,9 +1,6 @@
-use std::collections::HashMap;
-
 pub mod components;
+pub mod views;
 pub mod help;
 pub mod renderer;
 pub mod style;
 pub mod ui;
-
-pub type ViewStates = HashMap<String, String>;
