+++ src/view/views/logs/log_view.rs
