+++ src/view/views/mod.rs
@@ -1,5 +1,9 @@
+use std::collections::HashMap;
+
 pub mod app;
 pub mod logs;
 pub mod request;
 pub mod response;
 pub mod tabs_request;
+
+pub type ViewStates = HashMap<String, String>;
