+++ src/view/views/response/response_content_view.rs
